services/kafka-producer/requirements.txt
confluent-kafka==2.3.0
pyyaml==6.0.1
faker==22.6.0